% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balres}
\alias{balres}
\title{Datos del balance resumido de entidades financieras a partir del BCRA}
\format{
Un dataframe con 6 columnas:
\describe{
  \item{codigo_de_entidad}{Código de la entidad}
  \item{nombre_de_entidad}{Nombre de la entidad}
  \item{fecha}{Fecha}
  \item{codigo_de_linea}{Código de línea}
  \item{denominacion_de_la_cuenta}{Nombre de la cuenta}
  \item{valor}{Valor de la cuenta}
}
}
\source{
BCRA
}
\usage{
balres
}
\description{
Incluye información desde enero 2011
Nota: no se encontraron datos para 2011-2, 2011-6 y 2013-5
}
\keyword{datasets}
