% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balres_codigos.R
\name{balres_codigos}
\alias{balres_codigos}
\title{Lista de cuentas del balance resumido disponible en el txt Entfin/Tec_Cont/balres}
\usage{
balres_codigos()
}
\value{
Un dataframe con columnas: codigo_de_linea, denominacion_de_la_cuenta
}
\description{
Devuelve un dataframe con el código y nombre de las cuentas disponibles.
}
