% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entidades_codigos.R
\name{entidades_codigos}
\alias{entidades_codigos}
\title{Lista de entidades disponibles}
\usage{
entidades_codigos()
}
\value{
Un dataframe con columnas: codigo_entidad, nombre_de_entidad.
}
\description{
Devuelve un dataframe con el codigo y nombre de las entidades disponibles.
}
