% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fideicomisos_financieros_cnv}
\alias{fideicomisos_financieros_cnv}
\title{Datos de constitucion de fideicomisos financieros desde 2015 publicado por CNV.}
\format{
Un dataframe con 13 columnas:
\describe{
  \item{fecha_colocacion}{Fecha de colocacion}
  \item{fecha_de_liquidacion_y_emision}{Fecha de emision y liquidacion del fideicomiso}
  \item{fiduciario}{Nombre del fiduciario}
  \item{fiduciante}{Nombre del fiduciante}
  \item{denominacion_f_f}{Denominacion del fideicomiso financiero}
  \item{moneda}{Moneda}
  \item{monto_nominal_ff_moneda_emision}{Monto nominal de la emision}
  \item{categoria}{Categoria}
  \item{precio_de_corte_vrd_senior}{Precio de corte}
  \item{tipo_de_tasa}{Tipo de tasa}
  \item{tna_inicial_vrd_senior}{TNA inicial VRD senior}
  \item{tir_vrd_senior}{TIR VRD senior}
  \item{plazo_ff_meses}{Plazo en meses}
}
}
\source{
CNV
}
\usage{
fideicomisos_financieros_cnv
}
\description{
Datos de constitucion de fideicomisos financieros desde 2015 publicado por CNV.
}
\keyword{datasets}
