% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baldet.R
\name{get_baldet}
\alias{get_baldet}
\title{Actualiza los datos de balance detallado del BCRA}
\usage{
get_baldet(fecha_inicio = "2025-01-01", fecha_fin = Sys.Date())
}
\arguments{
\item{fecha_inicio}{Fecha de inicio de los datos (por defecto "2025-01-01")}

\item{fecha_fin}{Fecha del ultimo dato (por defecto fecha actual)}
}
\value{
Dataframe con los datos de balance detallado
}
\description{
Carga el balance detallado de todas las entidades para el periodo seleccionado desde la web del BCRA.
La carga puede demorar para periodos extensos debido al peso de cada archivo.
}
