% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniser.R
\name{get_uniser}
\alias{get_uniser}
\title{Obtiene el detalle de unidades de servicio por entidad}
\usage{
get_uniser(entidad = NULL, codigo_entidad = NULL)
}
\arguments{
\item{entidad}{Nombre (o parte del nombre) de la entidad (ej: "Galicia"). Por defecto trae todas.}

\item{codigo_entidad}{Codigo(s) exacto(s) de la entidad. Opcional.}
}
\value{
Un dataframe filtrado.
}
\description{
Permite filtrar por entidad para obtener sucursales, casas matrices y otros tipos de unidad por fecha de alta/baja y ubicacion geografica
}
