% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicadores_bancos}
\alias{indicadores_bancos}
\title{Datos de indicadores de entidades financieras a partir del BCRA}
\format{
Un dataframe con 6 columnas:
\describe{
  \item{codigo_de_entidad}{Código de la entidad}
  \item{nombre_de_entidad}{Nombre de la entidad}
  \item{fecha}{Fecha}
  \item{codigo_de_linea}{Código de línea}
  \item{descripcion_del_indicador}{descripción del indicador}
  \item{valor}{Valor del indicador}
}
}
\source{
BCRA
}
\usage{
indicadores_bancos
}
\description{
Incluye información desde enero 2011
Nota: no se encontraron datos para 2011-2, 2011-6 y 2013-5
}
\keyword{datasets}
