% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_codigos.R
\name{indicadores_codigos}
\alias{indicadores_codigos}
\title{Lista de indicadores financieros disponibles en el txt Entfin/Tec_Cont/indicad}
\usage{
indicadores_codigos()
}
\value{
Un dataframe con columnas: codigo_indicador, nombre_indicador.
}
\description{
Devuelve un dataframe con el codigo y nombre de los indicadores disponibles.
}
