% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{on_cnv}
\alias{on_cnv}
\title{Datos de emisiones de obligaciones negociables publicado por la CNV. Datos desde 2015.}
\format{
Un dataframe con 12 columnas:
\describe{
  \item{fecha_colocacion}{Fecha de colocacion de la ON}
  \item{fecha_de_emision_y_liquidacion}{Fecha de liquidacion de la ON}
  \item{sociedad}{Nombre de la sociedad emisora}
  \item{serie_clase}{Serie/clase de la ON}
  \item{moneda}{Moneda}
  \item{monto_nominal_moneda_emision}{Monto nominal de la emision}
  \item{regimen_de_emision}{Regimen de la emision}
  \item{tipo_de_tasa}{Tipo de tasa}
  \item{precio_de_corte}{Precio de corte}
  \item{tna_inicial}{TNA inicial}
  \item{tir_inicial}{TIR inicial}
  \item{plazo_meses}{Plazo en meses}
}
}
\source{
CNV
}
\usage{
on_cnv
}
\description{
Datos de emisiones de obligaciones negociables publicado por la CNV. Datos desde 2015.
}
\keyword{datasets}
