% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_balres.R
\name{update_balres}
\alias{update_balres}
\title{Actualiza los datos de balance resumido del BCRA}
\usage{
update_balres()
}
\value{
Nada. Guarda el archivo actualizado en `data/` del paquete.
}
\description{
Verifica que exista un archivo en la fecha siguiente a la ultima y lo agrega al
objeto `balres.rda`.
}
