% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cnv.R
\name{update_cnv}
\alias{update_cnv}
\title{Actualiza los datos de ON y Fideicomisos Financieros de CNV}
\usage{
update_cnv()
}
\value{
Nada. Guarda los archivos actualizados en `data/` del paquete.
}
\description{
Descarga los archivos de series estadísticas del sitio de CNV y actualiza
los objetos `on_cnv` y `fideicomisos_financieros_cnv`.
}
