% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Ejecuta todas las funciones de actualizacion de datos del paquete}
\usage{
update_data()
}
\value{
Nada. Ejecuta funciones con efecto de escritura en la carpeta `data/`.
}
\description{
Llama en orden a las funciones update_XXX() que descargan y procesan datos de las distintas fuentes.
}
