% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_indicadores.R
\name{update_indicadores}
\alias{update_indicadores}
\title{Actualiza los datos de indicadores financieros del BCRA}
\usage{
update_indicadores()
}
\value{
Nada. Guarda el archivo actualizado en `data/` del paquete.
}
\description{
Verifica que exista un archivo en la fecha siguiente a la última y lo agrega al
objeto `indicadores_bancos.rda`.
}
