% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_uniser.R
\name{update_uniser}
\alias{update_uniser}
\title{Actualiza los datos de unidades de servicio del BCRA}
\usage{
update_uniser()
}
\value{
Nada. Guarda el archivo actualizado en `data/` del paquete.
}
\description{
Verifica que exista un cambio en el excel uniser y, de existir, guarda la nueva
versión en el objeto `uniser.rda`.
}
