% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicadores.R
\name{get_indicadores}
\alias{get_indicadores}
\title{Obtiene indicadores de entidades por período y entidad}
\usage{
get_indicadores(
  periodos = NULL,
  entidad = NULL,
  codigo_entidad = NULL,
  codigo_indicador = NULL
)
}
\arguments{
\item{periodos}{Vector de períodos en formato "YYYY-MM". Si se pasan dos valores, se interpreta como rango.}

\item{entidad}{Nombre (o parte del nombre) de la entidad (ej: "Galicia"). Por defecto trae todas.}

\item{codigo_entidad}{Código(s) exacto(s) de la entidad. Opcional.}

\item{codigo_indicador}{Código(s) exacto(s) del indicador. Opcional.}
}
\value{
Un dataframe filtrado.
}
\description{
Filtra el dataset de indicadores financieros del BCRA por períodos seleccionados, entidad (por nombre o código) e indicador (por código).
}
