% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{on_cnv}
\alias{on_cnv}
\title{Datos de emisiones de obligaciones negociables publicado por la CNV. Datos desde 2015.}
\format{
Un dataframe con 12 columnas:
\describe{
  \item{fecha_colocacion}{Fecha de colocación de la ON}
  \item{fecha_de_emision_y_liquidacion}{Fecha de liquidación de la ON}
  \item{sociedad}{Nombre de la sociedad emisora}
  \item{serie_clase}{Serie/clase de la ON}
  \item{moneda}{Moneda}
  \item{monto_nominal_moneda_emision}{Monto nominal de la emisión}
  \item{regimen_de_emision}{Régimen de la emisión}
  \item{tipo_de_tasa}{Tipo de tasa}
  \item{precio_de_corte}{Precio de corte}
  \item{tna_inicial}{TNA inicial}
  \item{tir_inicial}{TIR inicial}
  \item{plazo_meses}{Plazo en meses}
}

Un dataframe con 13 columnas:
\describe{
  \item{fecha_colocacion}{Fecha de colocación}
  \item{fecha_de_liquidacion_y_emision}{Fecha de emisión y liquidación del fideicomiso}
  \item{fiduciario}{Nombre del fiduciario}
  \item{fiduciante}{Nombre del fiduciante}
  \item{denominacion_f_f}{Denominación del fideicomiso financiero}
  \item{moneda}{Moneda}
  \item{monto_nominal_ff_moneda_emision}{Monto nominal de la emisión}
  \item{categoria}{Categoría}
  \item{precio_de_corte_vrd_senior}{Precio de corte}
  \item{tipo_de_tasa}{Tipo de tasa}
  \item{tna_inicial_vrd_senior}{TNA inicial VRD senior}
  \item{tir_vrd_senior}{TIR VRD senior}
  \item{plazo_ff_meses}{Plazo en meses}
}
}
\source{
CNV

CNV
}
\usage{
on_cnv

on_cnv
}
\description{
Datos de emisiones de obligaciones negociables publicado por la CNV. Datos desde 2015.

Datos de constitución de fideicomisos financieros desde 2015 publicado por CNV.
}
\keyword{datasets}
