% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniser}
\alias{uniser}
\title{Datos de unidades de servicio por entidad de acuerdo a fechas de alta/baja y ubicación geográfica}
\format{
Un dataframe con 13 columnas:
\describe{
  \item{codigo_entidad}{Código de la entidad}
  \item{pais}{País de la unidad}
  \item{provincia}{Provincia de la unidad}
  \item{partido}{Partido o departamento de la unidad}
  \item{localidad}{Localidad de la unidad}
  \item{direccion}{Dirección de la unidad}
  \item{codigo_de_filial}{Código de filial}
  \item{tipo_de_unidad}{Tipo de unidad}
  \item{fecha_de_alta}{Fecha de alta de la unidad}
  \item{fecha_de_baja}{Fecha de baja de la unidad}
  \item{tipo_unidad}{Tipo de unidad}
  \item{nombre_entidad}{Nombre de la entidad}
  \item{grupo_entidad}{Grupo de la entidad}
}
}
\source{
BCRA
}
\usage{
uniser
}
\description{
Datos de unidades de servicio por entidad de acuerdo a fechas de alta/baja y ubicación geográfica
}
\keyword{datasets}
